refactor: Integrate stats module into math module
Integrate the new functionality into the math module instead
of creating a new module.
