+++ Makefile.am
@@ -74,8 +74,6 @@ test_bitmask_SOURCES = tests/test-bitmask.c tests/util.c
 test_bitmask_LDADD = libyara/.libs/libyara.a
 test_math_SOURCES = tests/test-math.c tests/util.c
 test_math_LDADD = libyara/.libs/libyara.a
-test_stats_SOURCES = tests/test-stats.c tests/util.c
-test_stats_LDADD = libyara/.libs/libyara.a
 test_stack_SOURCES = tests/test-stack.c tests/util.c
 test_stack_LDADD = libyara/.libs/libyara.a
 test_re_split_SOURCES = tests/test-re-split.c tests/util.c
@@ -97,7 +95,6 @@ check_PROGRAMS = \
   test-version \
   test-bitmask \
   test-math \
-  test-stats \
   test-stack \
   test-re-split \
   test-async
