+++ docs/modules/math.rst
@@ -118,3 +118,44 @@ file and create signatures based on those results.
     Returns 0 or 1, it's useful when writing a score based rule.
 
     *Example: math.to_number(SubRule1) \* 60 + math.to_number(SubRule2) \* 20 + math.to_number(SubRule3) \* 70 > 80*
+
+.. c:function:: count(byte/string, offset, size)
+
+    .. versionadded:: 4.2.0
+
+    Returns how often a specific byte or substring occurs, starting at *offset*
+    and looking at the next *size* bytes. When scanning a
+    running process the *offset* argument should be a virtual address within
+    the process address space.
+    *offset* and *size* are optional; if left empty, the complete file is searched.
+
+    *Example: stats.count("$[]", 0, 100) >= 5*
+
+    *Example: stats.count(0x4A) >= 10*
+
+.. c:function:: percentage(byte, offset, size)
+
+    .. versionadded:: 4.2.0
+
+    Returns the occurrence rate of a specific byte, starting at *offset*
+    and looking at the next *size* bytes. When scanning a
+    running process the *offset* argument should be a virtual address within
+    the process address space. The returned value is a float between 0 and 1.
+    *offset* and *size* are optional; if left empty, the complete file is searched.
+
+    
+    *Example: stats.percentage(0xFF, filesize-1024, filesize) >= 0.9*
+    
+    *Example: stats.percentage(0x4A) >= 0.4*
+
+.. c:function:: mode(offset, size)
+
+    .. versionadded:: 4.2.0
+    
+    Returns the most common byte, starting at *offset* and looking at the next
+    *size* bytes. When scanning a
+    running process the *offset* argument should be a virtual address within
+    the process address space. The returned value is a float.
+    *offset* and *size* are optional; if left empty, the complete file is searched.
+
+    *Example: stats.mode(0, filesize) == 0xFF*
