+++ tests/BUILD.bazel
@@ -172,17 +172,6 @@ cc_test(
 )
 
 cc_test(
-    name = "test_stats",
-    srcs = ["test-stats.c"],
-    copts = COPTS,
-    linkstatic = True,
-    deps = [
-        ":util",
-        "@//:libyara",
-    ],
-)
-
-cc_test(
     name = "test_stack",
     srcs = ["test-stack.c"],
     copts = COPTS,
