+++ tests/test-math.c
@@ -44,6 +44,86 @@ int main(int argc, char** argv)
       }",
       "A");
 
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.count(0x41, 0, 3) == 2 \
+      }",
+      "AABAAB");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.count(0x41) == 2 \
+      }",
+      "ABAB");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.count(\"AA\", 0, 10) == 4 \
+      }",
+      "AAAAA");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.count(\"ABC\", 0, 10) == 2 \
+      }",
+      "ABCDABCDABCD");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.count(\"ABC\") == 3 \
+      }",
+      "ABCDABCDABCD");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.count(\"AA\") == 4 \
+      }",
+      "AAAAA");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.percentage(0x41) > 0.39 and math.percentage(0x41) < 0.41 \
+      }",
+      "ABAB"); // Blob matching includes terminating zero byte
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.percentage(0x41, 0, 4) == 0.5 \
+      }",
+      "ABABCDEF");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.mode() == 0x41 \
+      }",
+      "ABABA");
+
+  assert_true_rule_blob(
+      "import \"math\" \
+      rule test { \
+        condition: \
+          math.mode(2, 3) == 0x41 \
+      }",
+      "CCABACC");
+
   yr_finalize();
 
   YR_DEBUG_FPRINTF(
